# api-tests/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Noah Explorer api",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/v1/blocks" : {
      "get" : {
        "tags" : [ "Blocks" ],
        "summary" : "Get list of blocks",
        "operationId" : "getBlocks",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Blocks per page",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of blocks",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BlockPaginatedCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blocks/{height}" : {
      "get" : {
        "tags" : [ "Blocks" ],
        "summary" : "Get block detail",
        "operationId" : "getBlock",
        "parameters" : [ {
          "name" : "height",
          "in" : "path",
          "description" : "Block height",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of blocks",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "404" : {
            "description" : "Block not found",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blocks/{height}/transactions" : {
      "get" : {
        "tags" : [ "Blocks" ],
        "summary" : "Get list of transactions by block",
        "operationId" : "getBlockTransactions",
        "parameters" : [ {
          "name" : "height",
          "in" : "path",
          "description" : "Block height",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of transactions",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionPaginatedCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/coins" : {
      "get" : {
        "tags" : [ "Coins" ],
        "summary" : "Get list of coins",
        "operationId" : "getCoins",
        "parameters" : [ {
          "name" : "symbol",
          "in" : "query",
          "description" : "Coin symbol for filter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of coins",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoinCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addresses" : {
      "get" : {
        "tags" : [ "Addresses" ],
        "summary" : "Get list of addresses",
        "operationId" : "getAddresses",
        "parameters" : [ {
          "name" : "addresses",
          "in" : "query",
          "description" : "List of addresses",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of Noah addresses",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AddressCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addresses/{address}" : {
      "get" : {
        "tags" : [ "Addresses" ],
        "summary" : "Get Noah address detail",
        "operationId" : "getAddress",
        "parameters" : [ {
          "name" : "address",
          "in" : "path",
          "description" : "Noah address",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Noah address details",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transactions" : {
      "get" : {
        "tags" : [ "Transactions" ],
        "summary" : "Get transactions list",
        "operationId" : "getTransactions",
        "parameters" : [ {
          "name" : "addresses",
          "in" : "query",
          "description" : "List of addresses",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "startblock",
          "in" : "query",
          "description" : "Start block id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "endblock",
          "in" : "query",
          "description" : "End block id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of transactions",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionPaginatedCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "BlockResponse" : {
        "required" : [ "blockTime", "hash", "height", "reward", "size", "timestamp", "txCount", "validators" ],
        "type" : "object",
        "properties" : {
          "height" : {
            "type" : "integer",
            "description" : "Block height",
            "example" : 34863
          },
          "size" : {
            "type" : "integer",
            "description" : "Block size",
            "example" : 253
          },
          "txCount" : {
            "type" : "integer",
            "description" : "Transaction count",
            "example" : 125
          },
          "blockTime" : {
            "type" : "number",
            "description" : "Block time",
            "example" : 4.23123
          },
          "timestamp" : {
            "type" : "string",
            "description" : "Block timestamp",
            "example" : "2019-02-19T00:00:00+03:00"
          },
          "reward" : {
            "type" : "string",
            "description" : "Block reward",
            "example" : "333"
          },
          "hash" : {
            "type" : "string",
            "description" : "Block hash",
            "example" : "ITHA0XQPS5XTOY6NTTOOJX5VAVZQHTRIZJDHQKKUPUQDNX7JYVDE4UO5LFFCWSVA"
          },
          "validators" : {
            "type" : "array",
            "description" : "Block validators",
            "items" : {
              "$ref" : "#/components/schemas/BlockValidatorResponse"
            }
          }
        }
      },
      "BlockPaginatedCollectionResponse" : {
        "required" : [ "data", "links", "meta" ],
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "description" : "Blocks",
            "items" : {
              "$ref" : "#/components/schemas/BlockResponse"
            }
          },
          "links" : {
            "$ref" : "#/components/schemas/PaginationLinksResponse"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMetaResponse"
          }
        }
      },
      "CoinCollectionResponse" : {
        "required" : [ "data" ],
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CoinResponse"
            }
          }
        }
      },
      "CoinResponse" : {
        "required" : [ "symbol" ],
        "type" : "object",
        "properties" : {
          "crr" : {
            "type" : "integer",
            "description" : "Coin crr",
            "example" : 10
          },
          "volume" : {
            "type" : "string",
            "description" : "Coin volume",
            "example" : "45656534.3546"
          },
          "reserve_balance" : {
            "type" : "string",
            "description" : "Coin reserve balance",
            "example" : "10.356"
          },
          "name" : {
            "type" : "string",
            "description" : "Coin name",
            "example" : "My super coin"
          },
          "symbol" : {
            "type" : "string",
            "description" : "Coin symbol",
            "example" : "MNT"
          }
        }
      },
      "PaginationLinksResponse" : {
        "required" : [ "first", "last", "next", "prev" ],
        "type" : "object",
        "properties" : {
          "first" : {
            "type" : "string",
            "description" : "First page",
            "example" : "https://testnet.explorer.noah.network/blocks?page=1"
          },
          "last" : {
            "type" : "string",
            "description" : "Last page",
            "example" : "https://testnet.explorer.noah.network/blocks?page=10"
          },
          "prev" : {
            "description" : "Previous page",
            "nullable" : true,
            "example" : "https://testnet.explorer.noah.network/blocks?page=1"
          },
          "next" : {
            "description" : "Next page",
            "nullable" : true,
            "example" : "https://testnet.explorer.noah.network/blocks?page=2"
          }
        }
      },
      "PaginationMetaResponse" : {
        "required" : [ "current_page", "last_page", "path", "per_page", "total" ],
        "type" : "object",
        "properties" : {
          "current_page" : {
            "type" : "integer",
            "description" : "Current page number",
            "example" : 2
          },
          "last_page" : {
            "type" : "integer",
            "description" : "Last page number",
            "example" : 10
          },
          "path" : {
            "type" : "string",
            "description" : "Request path",
            "example" : "https://testnet.explorer.noah.network"
          },
          "per_page" : {
            "type" : "integer",
            "description" : "Elements per page",
            "example" : 50
          },
          "total" : {
            "type" : "integer",
            "description" : "Total number of elements",
            "example" : 16434
          }
        }
      },
      "AddressCollectionResponse" : {
        "required" : [ "data" ],
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AddressResponse"
            }
          }
        }
      },
      "AddressResponse" : {
        "required" : [ "address", "balances" ],
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string",
            "description" : "Noah address",
            "example" : "Noahxe449e9fd85fd553fa45d245b62e30253bce29cc7"
          },
          "balances" : {
            "type" : "array",
            "description" : "Address balances",
            "items" : {
              "$ref" : "#/components/schemas/AddressBalanceResponse"
            }
          }
        }
      },
      "AddressBalanceResponse" : {
        "required" : [ "amount", "coin" ],
        "properties" : {
          "coin" : {
            "type" : "string",
            "description" : "Coin symbol",
            "example" : "MNT"
          },
          "amount" : {
            "type" : "string",
            "description" : "Amount of coin",
            "example" : "1257.4657"
          }
        }
      },
      "TransactionResponse" : {
        "required" : [ "block", "data", "fee", "from", "hash", "nonce", "payload", "timestamp", "txn", "type" ],
        "properties" : {
          "txn" : {
            "type" : "integer",
            "description" : "Transaction number",
            "example" : 17653
          },
          "hash" : {
            "type" : "string",
            "description" : "Transaction hash",
            "example" : "Mtc923c18c1d842dcba262c63df22003d5ee4e683dc2927acf9580f083ad9e193e"
          },
          "nonce" : {
            "type" : "integer",
            "description" : "Transaction nonce",
            "example" : 113
          },
          "block" : {
            "type" : "integer",
            "description" : "Block id",
            "example" : 119
          },
          "timestamp" : {
            "type" : "string",
            "description" : "Transaction timestamp",
            "example" : "2019-02-19T00:00:00+03:00"
          },
          "fee" : {
            "type" : "string",
            "description" : "Transaction fee",
            "example" : "10.000000000000000000"
          },
          "type" : {
            "type" : "integer",
            "description" : "Transaction type",
            "example" : 1
          },
          "payload" : {
            "type" : "string",
            "description" : "Transaction payload",
            "example" : "Super tx from developers"
          },
          "from" : {
            "type" : "string",
            "description" : "Transaction sender address",
            "example" : "Noahx184ac726059e43643e67290666f7b3195093f870"
          },
          "data" : {
            "$ref" : "#/components/schemas/TxSendCoinData"
          }
        }
      },
...
      "TransactionPaginatedCollectionResponse" : {
        "required" : [ "data", "links", "meta" ],
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "description" : "Transactions",
            "items" : {
              "$ref" : "#/components/schemas/TransactionResponse"
            }
          },
          "links" : {
            "$ref" : "#/components/schemas/PaginationLinksResponse"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMetaResponse"
          }
        }
      },
      "BlockValidatorResponse" : {
        "required" : [ "publicKey", "signed" ],
        "properties" : {
          "publicKey" : {
            "type" : "string",
            "description" : "Validator public key",
            "example" : "Mp8f053f3802d33f5e7092bb01ca99ae77606f4faf759c72560d5ee69b8e191a56"
          },
          "signed" : {
            "type" : "boolean",
            "description" : "Is block signed by validator",
            "example" : true
          }
        }
      },
...
      "TxSendCoinData" : {
        "properties" : {
          "coin" : {
            "type" : "string",
            "description" : "Send coin",
            "example" : "MNT"
          },
          "to" : {
            "type" : "string",
            "description" : "Receiver address",
            "example" : "Noahxee81347211c72524338f9680072af90744333146"
          },
          "value" : {
            "type" : "string",
            "description" : "Send coin amount",
            "example" : "12.357"
          }
        }
      },
...
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BlockResponse"
          }
        }
      },
      "inline_response_404_error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "example" : 404
          },
          "message" : {
            "type" : "string",
            "example" : "Block not found."
          }
        }
      },
      "inline_response_404" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/inline_response_404_error"
          }
        }
      },
      "inline_response_200_1" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/AddressResponse"
          }
        }
      },
...
    }
  }
}